pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["MarinaIlina893 <marinailina893@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest-cov = "^3.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^6.2.5"
ipython = "^7.28.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
